Rename people types to Donor and Donors, other improvements
An -email flag can be provided to just print emails of donors. I used
this to get a list of all emails for donors in 2019 and 2020 to ask if
they wished to be publicly acknowledged for their donation or wanted to
stay anonymous.

The feedback from this was used to add a few new features. A new
donors.json config file can be provided to list emails of donors who
wish to remain anonymous and also to correct some names based on what
some donors preferred to be called. This is currently just used when the
'donors' command is run but will also be used for future work to
generate the donor thanks page.